Improved debugging experience.
* Update Roslyn to 2.3.1.
* Use standardized way of emitting debug symbols for debug build.
* Unify script sessions.
