+++ src/Cake/Cake.csproj
@@ -19,7 +19,7 @@
 
   <!-- Global packages -->
   <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="2.1.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Scripting" Version="2.3.1" />
   </ItemGroup>
 
   <!-- .NET Core packages -->
