--- src/Cake/Scripting/Roslyn/RoslynScriptSessionBase.cs
@@ -1,72 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Collections.Generic;
-using System.Reflection;
-using Cake.Core.Diagnostics;
-using Cake.Core.IO;
-using Cake.Core.Reflection;
-using Cake.Core.Scripting;
-
-namespace Cake.Scripting.Roslyn
-{
-    internal abstract class RoslynScriptSessionBase : IScriptSession
-    {
-        private readonly IAssemblyLoader _loader;
-        private readonly ICakeLog _log;
-
-        public HashSet<FilePath> ReferencePaths { get; }
-
-        public HashSet<Assembly> References { get; }
-
-        public HashSet<string> Namespaces { get; }
-
-        protected RoslynScriptSessionBase(IAssemblyLoader loader, ICakeLog log)
-        {
-            _loader = loader;
-            _log = log;
-
-            ReferencePaths = new HashSet<FilePath>(PathComparer.Default);
-            References = new HashSet<Assembly>();
-            Namespaces = new HashSet<string>(StringComparer.Ordinal);
-        }
-
-        public void AddReference(Assembly assembly)
-        {
-            if (assembly == null)
-            {
-                throw new ArgumentNullException(nameof(assembly));
-            }
-            _log.Debug("Adding assembly reference to {0}...", new FilePath(assembly.Location).GetFilename().FullPath);
-            References.Add(assembly);
-        }
-
-        public void AddReference(FilePath path)
-        {
-            if (path == null)
-            {
-                throw new ArgumentNullException(nameof(path));
-            }
-
-            _log.Debug("Adding reference to {0}...", path.GetFilename().FullPath);
-#if NETCORE
-            References.Add(_loader.Load(path, true));
-#else
-            ReferencePaths.Add(path);
-#endif
-        }
-
-        public void ImportNamespace(string @namespace)
-        {
-            if (!Namespaces.Contains(@namespace))
-            {
-                _log.Debug("Importing namespace {0}...", @namespace);
-                Namespaces.Add(@namespace);
-            }
-        }
-
-        public abstract void Execute(Script script);
-    }
-}
